Replaced float with double and converted float literals with doubles.

+++ Invicta.Numerics.Vectors.Tests/Numerics/Tests/PlaneDTests.cs
@@ -13,8 +13,8 @@ namespace Invicta.Numerics.Tests
         [Fact]
         public void PlaneDEqualsTest1()
         {
-            PlaneD a = new PlaneD(1.0f, 2.0f, 3.0f, 4.0f);
-            PlaneD b = new PlaneD(1.0f, 2.0f, 3.0f, 4.0f);
+            PlaneD a = new PlaneD(1.0d, 2.0d, 3.0d, 4.0d);
+            PlaneD b = new PlaneD(1.0d, 2.0d, 3.0d, 4.0d);
 
             // case 1: compare between same values
             bool expected = true;
@@ -22,7 +22,7 @@ namespace Invicta.Numerics.Tests
             Assert.Equal(expected, actual);
 
             // case 2: compare between different values
-            b.Normal = new Vector3D(10.0f, b.Normal.Y, b.Normal.Z);
+            b.Normal = new Vector3D(10.0d, b.Normal.Y, b.Normal.Z);
             expected = false;
             actual = a.Equals(b);
             Assert.Equal(expected, actual);
@@ -32,8 +32,8 @@ namespace Invicta.Numerics.Tests
         [Fact]
         public void PlaneDEqualsTest()
         {
-            PlaneD a = new PlaneD(1.0f, 2.0f, 3.0f, 4.0f);
-            PlaneD b = new PlaneD(1.0f, 2.0f, 3.0f, 4.0f);
+            PlaneD a = new PlaneD(1.0d, 2.0d, 3.0d, 4.0d);
+            PlaneD b = new PlaneD(1.0d, 2.0d, 3.0d, 4.0d);
 
             // case 1: compare between same values
             object obj = b;
@@ -43,7 +43,7 @@ namespace Invicta.Numerics.Tests
             Assert.Equal(expected, actual);
 
             // case 2: compare between different values
-            b.Normal = new Vector3D(10.0f, b.Normal.Y, b.Normal.Z);
+            b.Normal = new Vector3D(10.0d, b.Normal.Y, b.Normal.Z);
 
             obj = b;
             expected = false;
@@ -67,8 +67,8 @@ namespace Invicta.Numerics.Tests
         [Fact]
         public void PlaneDInequalityTest()
         {
-            PlaneD a = new PlaneD(1.0f, 2.0f, 3.0f, 4.0f);
-            PlaneD b = new PlaneD(1.0f, 2.0f, 3.0f, 4.0f);
+            PlaneD a = new PlaneD(1.0d, 2.0d, 3.0d, 4.0d);
+            PlaneD b = new PlaneD(1.0d, 2.0d, 3.0d, 4.0d);
 
             // case 1: compare between same values
             bool expected = false;
@@ -76,7 +76,7 @@ namespace Invicta.Numerics.Tests
             Assert.Equal(expected, actual);
 
             // case 2: compare between different values
-            b.Normal = new Vector3D(10.0f, b.Normal.Y, b.Normal.Z);
+            b.Normal = new Vector3D(10.0d, b.Normal.Y, b.Normal.Z);
             expected = true;
             actual = a != b;
             Assert.Equal(expected, actual);
@@ -86,8 +86,8 @@ namespace Invicta.Numerics.Tests
         [Fact]
         public void PlaneDEqualityTest()
         {
-            PlaneD a = new PlaneD(1.0f, 2.0f, 3.0f, 4.0f);
-            PlaneD b = new PlaneD(1.0f, 2.0f, 3.0f, 4.0f);
+            PlaneD a = new PlaneD(1.0d, 2.0d, 3.0d, 4.0d);
+            PlaneD b = new PlaneD(1.0d, 2.0d, 3.0d, 4.0d);
 
             // case 1: compare between same values
             bool expected = true;
@@ -95,7 +95,7 @@ namespace Invicta.Numerics.Tests
             Assert.Equal(expected, actual);
 
             // case 2: compare between different values
-            b.Normal = new Vector3D(10.0f, b.Normal.Y, b.Normal.Z);
+            b.Normal = new Vector3D(10.0d, b.Normal.Y, b.Normal.Z);
             expected = false;
             actual = a == b;
             Assert.Equal(expected, actual);
@@ -105,18 +105,18 @@ namespace Invicta.Numerics.Tests
         [Fact]
         public void PlaneDGetHashCodeTest()
         {
-            PlaneD target = new PlaneD(1.0f, 2.0f, 3.0f, 4.0f);
+            PlaneD target = new PlaneD(1.0d, 2.0d, 3.0d, 4.0d);
 
             int expected = HashCode.Combine(target.Normal, target.D);
             int actual = target.GetHashCode();
             Assert.Equal(expected, actual);
         }
 
-        // A test for PlaneD (float, float, float, float)
+        // A test for PlaneD (double, double, double, double)
         [Fact]
         public void PlaneDConstructorTest1()
         {
-            float a = 1.0f, b = 2.0f, c = 3.0f, d = 4.0f;
+            double a = 1.0d, b = 2.0d, c = 3.0d, d = 4.0d;
             PlaneD target = new PlaneD(a, b, c, d);
 
             Assert.True(
@@ -128,12 +128,12 @@ namespace Invicta.Numerics.Tests
         [Fact]
         public void PlaneDCreateFromVerticesTest()
         {
-            Vector3D point1 = new Vector3D(0.0f, 1.0f, 1.0f);
-            Vector3D point2 = new Vector3D(0.0f, 0.0f, 1.0f);
-            Vector3D point3 = new Vector3D(1.0f, 0.0f, 1.0f);
+            Vector3D point1 = new Vector3D(0.0d, 1.0d, 1.0d);
+            Vector3D point2 = new Vector3D(0.0d, 0.0d, 1.0d);
+            Vector3D point3 = new Vector3D(1.0d, 0.0d, 1.0d);
 
             PlaneD target = PlaneD.CreateFromVertices(point1, point2, point3);
-            PlaneD expected = new PlaneD(new Vector3D(0, 0, 1), -1.0f);
+            PlaneD expected = new PlaneD(new Vector3D(0, 0, 1), -1.0d);
             Assert.Equal(target, expected);
         }
 
@@ -141,23 +141,23 @@ namespace Invicta.Numerics.Tests
         [Fact]
         public void PlaneDCreateFromVerticesTest2()
         {
-            Vector3D point1 = new Vector3D(0.0f, 0.0f, 1.0f);
-            Vector3D point2 = new Vector3D(1.0f, 0.0f, 0.0f);
-            Vector3D point3 = new Vector3D(1.0f, 1.0f, 0.0f);
+            Vector3D point1 = new Vector3D(0.0d, 0.0d, 1.0d);
+            Vector3D point2 = new Vector3D(1.0d, 0.0d, 0.0d);
+            Vector3D point3 = new Vector3D(1.0d, 1.0d, 0.0d);
 
             PlaneD target = PlaneD.CreateFromVertices(point1, point2, point3);
-            float invRoot2 = (float)(1 / Math.Sqrt(2));
+            double invRoot2 = (double)(1 / Math.Sqrt(2));
 
             PlaneD expected = new PlaneD(new Vector3D(invRoot2, 0, invRoot2), -invRoot2);
             Assert.True(MathHelper.Equal(target, expected), "PlaneD.cstor did not return the expected value.");
         }
 
-        // A test for PlaneD (Vector3Df, float)
+        // A test for PlaneD (Vector3Df, double)
         [Fact]
         public void PlaneDConstructorTest3()
         {
             Vector3D normal = new Vector3D(1, 2, 3);
-            float d = 4;
+            double d = 4;
 
             PlaneD target = new PlaneD(normal, d);
             Assert.True(
@@ -169,7 +169,7 @@ namespace Invicta.Numerics.Tests
         [Fact]
         public void PlaneDConstructorTest()
         {
-            Vector4D value = new Vector4D(1.0f, 2.0f, 3.0f, 4.0f);
+            Vector4D value = new Vector4D(1.0d, 2.0d, 3.0d, 4.0d);
             PlaneD target = new PlaneD(value);
 
             Assert.True(
@@ -183,8 +183,8 @@ namespace Invicta.Numerics.Tests
             PlaneD target = new PlaneD(2, 3, 4, 5);
             Vector4D value = new Vector4D(5, 4, 3, 2);
 
-            float expected = 10 + 12 + 12 + 10;
-            float actual = PlaneD.Dot(target, value);
+            double expected = 10 + 12 + 12 + 10;
+            double actual = PlaneD.Dot(target, value);
             Assert.True(MathHelper.Equal(expected, actual), "PlaneD.Dot returns unexpected value.");
         }
 
@@ -194,8 +194,8 @@ namespace Invicta.Numerics.Tests
             PlaneD target = new PlaneD(2, 3, 4, 5);
             Vector3D value = new Vector3D(5, 4, 3);
 
-            float expected = 10 + 12 + 12 + 5;
-            float actual = PlaneD.DotCoordinate(target, value);
+            double expected = 10 + 12 + 12 + 5;
+            double actual = PlaneD.DotCoordinate(target, value);
             Assert.True(MathHelper.Equal(expected, actual), "PlaneD.DotCoordinate returns unexpected value.");
         }
 
@@ -205,8 +205,8 @@ namespace Invicta.Numerics.Tests
             PlaneD target = new PlaneD(2, 3, 4, 5);
             Vector3D value = new Vector3D(5, 4, 3);
 
-            float expected = 10 + 12 + 12;
-            float actual = PlaneD.DotNormal(target, value);
+            double expected = 10 + 12 + 12;
+            double actual = PlaneD.DotNormal(target, value);
             Assert.True(MathHelper.Equal(expected, actual), "PlaneD.DotCoordinate returns unexpected value.");
         }
 
@@ -215,8 +215,8 @@ namespace Invicta.Numerics.Tests
         {
             PlaneD target = new PlaneD(1, 2, 3, 4);
 
-            float f = target.Normal.LengthSquared();
-            float invF = 1.0f / (float)Math.Sqrt(f);
+            double f = target.Normal.LengthSquared();
+            double invF = 1.0d / (double)Math.Sqrt(f);
             PlaneD expected = new PlaneD(target.Normal * invF, target.D * invF);
 
             PlaneD actual = PlaneD.Normalize(target);
@@ -235,18 +235,18 @@ namespace Invicta.Numerics.Tests
             target = PlaneD.Normalize(target);
 
             Matrix4x4D m =
-                Matrix4x4D.CreateRotationX(MathHelper.ToRadians(30.0f)) *
-                Matrix4x4D.CreateRotationY(MathHelper.ToRadians(30.0f)) *
-                Matrix4x4D.CreateRotationZ(MathHelper.ToRadians(30.0f));
-            m.M41 = 10.0f;
-            m.M42 = 20.0f;
-            m.M43 = 30.0f;
+                Matrix4x4D.CreateRotationX(MathHelper.ToRadians(30.0d)) *
+                Matrix4x4D.CreateRotationY(MathHelper.ToRadians(30.0d)) *
+                Matrix4x4D.CreateRotationZ(MathHelper.ToRadians(30.0d));
+            m.M41 = 10.0d;
+            m.M42 = 20.0d;
+            m.M43 = 30.0d;
 
             PlaneD expected = new PlaneD();
             Matrix4x4D inv;
             Matrix4x4D.Invert(m, out inv);
             Matrix4x4D itm = Matrix4x4D.Transpose(inv);
-            float x = target.Normal.X, y = target.Normal.Y, z = target.Normal.Z, w = target.D;
+            double x = target.Normal.X, y = target.Normal.Y, z = target.Normal.Z, w = target.D;
             expected.Normal = new Vector3D(
                 x * itm.M11 + y * itm.M21 + z * itm.M31 + w * itm.M41,
                 x * itm.M12 + y * itm.M22 + z * itm.M32 + w * itm.M42,
@@ -266,13 +266,13 @@ namespace Invicta.Numerics.Tests
             target = PlaneD.Normalize(target);
 
             Matrix4x4D m =
-                Matrix4x4D.CreateRotationX(MathHelper.ToRadians(30.0f)) *
-                Matrix4x4D.CreateRotationY(MathHelper.ToRadians(30.0f)) *
-                Matrix4x4D.CreateRotationZ(MathHelper.ToRadians(30.0f));
+                Matrix4x4D.CreateRotationX(MathHelper.ToRadians(30.0d)) *
+                Matrix4x4D.CreateRotationY(MathHelper.ToRadians(30.0d)) *
+                Matrix4x4D.CreateRotationZ(MathHelper.ToRadians(30.0d));
             QuaternionD q = QuaternionD.CreateFromRotationMatrix(m);
 
             PlaneD expected = new PlaneD();
-            float x = target.Normal.X, y = target.Normal.Y, z = target.Normal.Z, w = target.D;
+            double x = target.Normal.X, y = target.Normal.Y, z = target.Normal.Z, w = target.D;
             expected.Normal = new Vector3D(
                 x * m.M11 + y * m.M21 + z * m.M31 + w * m.M41,
                 x * m.M12 + y * m.M22 + z * m.M32 + w * m.M42,
@@ -288,10 +288,10 @@ namespace Invicta.Numerics.Tests
         [Fact]
         public void PlaneDEqualsNaNTest()
         {
-            PlaneD a = new PlaneD(float.NaN, 0, 0, 0);
-            PlaneD b = new PlaneD(0, float.NaN, 0, 0);
-            PlaneD c = new PlaneD(0, 0, float.NaN, 0);
-            PlaneD d = new PlaneD(0, 0, 0, float.NaN);
+            PlaneD a = new PlaneD(double.NaN, 0, 0, 0);
+            PlaneD b = new PlaneD(0, double.NaN, 0, 0);
+            PlaneD c = new PlaneD(0, 0, double.NaN, 0);
+            PlaneD d = new PlaneD(0, 0, 0, double.NaN);
 
             Assert.False(a == new PlaneD(0, 0, 0, 0));
             Assert.False(b == new PlaneD(0, 0, 0, 0));
@@ -350,7 +350,7 @@ namespace Invicta.Numerics.Tests
         struct PlaneDPlusFloat
         {
             private PlaneD _v;
-            private float _f;
+            private double _f;
         }
 
         [StructLayout(LayoutKind.Sequential)]
@@ -366,7 +366,7 @@ namespace Invicta.Numerics.Tests
         {
             PlaneD plane = new PlaneD();
 
-            float* basePtr = &plane.Normal.X; // Take address of first element
+            double* basePtr = &plane.Normal.X; // Take address of first element
             PlaneD* planePtr = &plane; // Take address of whole PlaneD
 
             Assert.Equal(new IntPtr(basePtr), new IntPtr(planePtr));
